refactoring custId to long

+++ src/main/java/com/starboard/b2b/dao/BrandDao.java
@@ -10,6 +10,6 @@ public interface BrandDao {
 
 	List<Brand> list();
 
-	void addSelectedBrand(Integer custId, List<Integer> selectedBrand);
+	void addSelectedBrand(Long custId, List<Integer> selectedBrand);
 
 }
